Fix export of C++ extensions (#304)
Fixes #303
Revert changes to CustomComponent declaration to avoid breaking
extensions. Extensions that want to do initialization in the header like
the built-in objects can override the declaration field.
Restore export of construction field in subsystem. This is not used in
built-in objects but may be needed for extensions.
Fix a few old references to shared pointers